raspberry-pi/requirements.txt
selenium
python-dotenv
boto3
urllib3